+++ src/modules/CanvasRotate/CanvasRotate.hpp
@@ -15,6 +15,7 @@ class $editorModule(CanvasRotate) {
     void onEditor() override;
     void toggleBetterEditHook(bool enabled);
     bool isLassoActive();
+    bool isRotating();
 
     bool onToggled(bool state) override;
     bool onSettingChanged(std::string_view key, const matjson::Value& value) override;